added basic unit tests
